Update patch set 1
Patch Set 1:

This change LGTM (for xperms, not ops).

The general workflow for external/selinux is to submit changes to the upstream selinux project. We periodically do a git pull on the upstream git project to merge new changes in. If you want to do a review in gerrit before submitting changes upstream (I typically do), feel free to continue uploading changes here.

Patch-set: 1