Update patch set 1
Patch Set 1:

> Where did you see the build breakage? Nothing AFAIK has changed
 > here...

it was relying on bionic exposing a "machine/endian.h" file (but only for Mac builds) so checkbuild and treehugger were happy when we removed it, but the *device* build broke if built on a Mac. (so, no, unless you have a Mac, you can't reproduce the failure. death to the Mac build!)

Patch-set: 1