Update patch set 1
Patch Set 1:

libsepol is built as a dependency for checkpolicy, a host-side tool for building the policy.
If the issue is that machine/endian.h is now private to bionic, then how does this change fix that problem since we will still try to include it even after this change?
Why is this any different than the same usage in checkpolicy/Android.mk?
Is the correct fix to just replace #ifdef DARWIN with #ifdef _APPLE_ throughout libsepol and checkpolicy?

Patch-set: 1