Update patch set 1
Patch Set 1:

This configuration currently does work and libselinux is correctly indicated with the NSA's public domain license. However, bear in mind only the libselinux compoenents get installed to the device, so this may not work if another sub-component ever goes to the target.

Ill leave this up to others on how they want this managed, where each component has a notice file, or global + specific (current).

Patch-set: 1
Label: Verified=0