Update patch set 4
Patch Set 4:

> This should also go, as property contexts backend was upstreamed,
> and the same backend covers service_contexts.

Yeah about that. We figured that using the property contexts backend for the services was not such a good idea because service names should match exactly and properties are only matched by prefix. So there exists a patch in the pipeline that introduces a new backend based on the property backend, that matches exactly. I'll post a patch to upstream in a few minutes.

Patch-set: 4
Label: Verified=0