libsepol: cil: Add cil_attributize.
Add a cil_attributize() function which transforms a given policy into an
"attributized", versioned policy.  Specifically add a component which
traverses the cil AST to extract the objects to be "attributized",
currently just types and attributes, and add another component which
transforms the objects in a combined platform + non-platform policy into
attributized
versions if they match those found in the first pass.

Test: Tested by creating two example CIL policy files, one considered
'public platform' policy and the other a combination of the 'public
platform policy' and private, non-platform policy, and then verifying
that the combined policy elements which were in the 'public platform'
policy were changed to attributes with a version tag.

Bug: 31369363
Change-Id: I62f4b4ebacd6d125beb456fba0cd5024219949e7