Update patch set 1
Patch Set 1:

The version_map and attributize files could be combined once I fix Bug: 31594192, which was filed because the current setup of the global cil_strpool prevents having two cil_db objects active concurrently.

The code here would be used to operate on an exported platform policy and a one combined with non-platform private policy  

The next planned step is to take the platform policy, which I don't believe needs to be attributized in this manner, a mapping between platform types and old public platform policy versions, and the output of this  code to  combine them all into one final CIL representation which can be compiled down to the policy binary format.

Patch-set: 1