libsepol: cil: Add android.h
Add a cil_attributize() function which transforms a given policy into an
"attributized", versioned policy.  Specifically add a component which
traverses the cil AST to extract the objects to be "attributized",
currently just types and attributes, and add another component which
transforms the objects in a combined platform + non-platform policy into
attributized versions if they match those found in the first pass.

Also add a cil_attrib_mapping() function which will generate the mapping
to be included and modified in future platform versions to ensure
backwards compatibility between types and the resulting attributes.

Test: Tested by
1) creating two example CIL policy files, one considered
'public platform' policy and the other a combination of the 'public
platform policy' and private, non-platform policy, and then verifying
that the combined policy elements which were in the 'public platform'
policy were changed to attributes with a version tag

2) verifying that the changed types and attributes are produced in the
mapping file

3) by using secilc to compile the platform, versioned non-platform and
mapping files into a functioning policy binary (which runs on angler).

Bug: 31369363
Change-Id: I62f4b4ebacd6d125beb456fba0cd5024219949e7