libsepol: cil: cil_strpool: Allow multiple strpool users.
cil_strpool currently provides an interface to a statically stored
global data structure.  This interface does not accomodate multiple
consumers, however, as two calls to cil_strpool_init() will lead to a
memory leak and a call to cil_strpool_destroy() by one consumer will
remove data from use by others, and subsequently lead to a segfault on
the next cil_strpool_destroy() invocation.

Add a reference counter so that the strpool is only initialized once and
protect the exported interface with a mutex.

Test: Tested by calling cil_db_init() on two cil_dbs and then calling
cil_db_destroy() on each.

Bug: 31594192
Change-Id: Ibdf081fe2f0cb63b84e4585887898d059ccc28d4