Update patch set 2
Patch Set 2:

I am fine with the approach. Eventually, it would be better to get rid of the global state, but that is not something I want to burden you with.

Steve tells me that Red Hat had a nice trick to avoid the overhead of pthreads when possible, see: libselinux/src/selinux_internal.h

Patch-set: 2