Update patch set 1
Patch Set 1:

@Stephen, @James:  Any inclination as to whether or not this would be desirable upstream?

There are other ways to do this; I figured directly opening and loading the files would probably be the fastest, and nnk@ thought it would be conceptually easier to just pass in two paths, but I'd like to keep us in sync with upstream if possible.

This also has the drawback that an fc_sort() type operation may be needed on-device as well, although I think that may be unavoidable (this matches current behavior as-is).

Patch-set: 1
Label: Verified=0