Update patch set 1
Patch Set 1:

> (4 comments)

The reason I use __FILE__ and __LINE__ is so we know where the unsupported code lives...

I would not consider it an odd macro function because it needs a ';', that's pretty normal. Folks will often do do { xxx } while(0).

You're macro doesn't work, this macro function has a static return code of ENOTSUP, so we want that in front so it can be assigned to rc.

tabs/spaces not a problem, editor change :-P

Patch-set: 1