CIL: Add annotations to preserve or expand attributes
This commit adds two mutually exlusive options to the policy language
for type attributes. These options are applied directly to the
attribute declaration. This change has no impact on unannotated
attributes.

preserve:  Prevents an attribute from being removed from policy.

expand:    Always expands an attribute.

Example:
attribute foo expand;

CIL example:
(typeattribute hal_tv_cec_client expand)

Adding the annotations directly to policy was chosen over other
methods as it is consistent with how targeted runtime optimizations
are specified in other languages. For example, in C the "inline"
command. [1]

Motivation

Preserve:
Commit 0be23c3f15fd added the capability to aggresively remove unused
attributes. This is generally useful as too many attributes assigned
to a type results in lengthy policy look up times when there is a
cache miss. However, removing attributes can also result in loss of
information used in external tests. On Android, neverallow rules are
asserted at build time and during device certification (CTS), but are
stripped from policy files compiled on device. Because neverallow
rules are absent when secilc is run, some attributes are being
stripped from policy and neverallow tests in CTS may be violated. [2]
This change retains the aggressive attribute stripping behavior but
adds an override mechanism to preserve attributes marked as
necessary.

Expand:
Android has been moving away from a monolithic policy binary to
a two part split policy representing the Android platform and the
underlying vendor-provided hardware interface. The goal is a stable
API allowing these two parts to be updated independently of each
other. Attributes provide an important mechanism for compatibility.
For example, when the vendor provides a HAL for the platform,
permissions needed by clients of the HAL can be granted to an
attribute. Clients need only be assigned the attribute and do not
need to be aware of the underlying types and permissions being
granted.

Inheriting permissions via attribute creates a convenient mechanism
for independence between vendor and platform policy, but results
in the creation of many attributes, and the potential for performance
issues when processes are clients of many HALs. Annotating these
attributes for expansion at compile time allows us to retain the
compatibility benefits of using attributes without the performance
costs. [3]

[1] Also considered passing in a list of attributes to expand or
    preserve as command line options to secilc.
[2] data_file_type and service_manager_type are stripped from AOSP
    policy when using secilc's -G option. This impacts 11 neverallow
    tests in CTS.
[3] Annotating all HAL client attributes for expansion resulted in
    system_server's dropping from 19 attributes to 8. Because these
    attributes were not widely applied to other types, the final
    policy size change was negligible.

Bug: 37357742
Bug: 36508258
Test: Build and boot Marlin with all hal_*_client attributes marked
      for expansion. Verify (using seinfo and sesearch) that
      permissions are correctly expanded from attributes to types.
Test: Mark types being stripped by secilc with "preserve" and verify
      that they are retained in policy and applied to the same types.

Change-Id: I14d99662106baa643ea3e0375390a11867e9cd81
Signed-off-by: Jeff Vander Stoep <jeffv@google.com>