cil: allow typepermissive on platform public types
Currently, adding a permissive statement on a platform public type
is not allowed, and throws the following build error:
typepermissive contains platform public type: kernel (line 879) .

This is problematic, when you have a single domain preventing
enabling the global enforcement mode. It would be better to
keep the type permissive on that single domain and have enforcement
turned on for others. Additionally, the permissive domain check is
run on user build variants, so some sanity checking of it's usage
occurs, including during CTS. Kill this check and allow it.

Test: add a permissive kernel to device policy and build, boot hikey.

Change-Id: Idc84ace2129dd0d9ec8726f8d80044836acd4377
Signed-off-by: William Roberts <william.c.roberts@intel.com>