Update patch set 1
Patch Set 1:

Can we easily survey the property_contexts in use on a range of OEM devices (in addition what we can already see in AOSP device/*/* trees and system/sepolicy), to get a sense of whether this would be a significant compatibility problem?
I'd be inclined to say that we just change the default and document it for the next Android release.  Upstream backward compat isn't really an issue here because this config is only used in Android.  But I'm fine either way.

Patch-set: 1
CC: Gerrit User 1010111 <1010111@85c56323-6fa9-3386-8a01-6480fb634889>