Update patch set 1
Patch Set 1:

> > > I do think there's some merit to changing the syntax if it
 > isn't
 > > a
 > > > substantial amount of work.  I'd like to hear from Jaekyun what
 > > he
 > > > thinks.
 > > >
 > > > '$' will never appear in system properties (https://android.googlesource.com/platform/system/core/+/master/init/property_service.cpp#125)
 > > > and I recently closed out a discussion about expanding that
 > list
 > > of
 > > > symbols with the conclusion that we should keep it is, so we
 > > should
 > > > be good if we choose that route.
 > >
 > >
 > >
 > >
 > >
 > > I missed those entries, I actually looked :-(
 > >
 > > Why not just make it like seapp_contexts and require the * for
 > > prefix matching at the end, this way it's explicit what you want.
 > 
 > I think the compatibility argument has more merit given that AOSP
 > itself has relied upon the current prefix matching behavior since
 > the beginning.  I wouldn't want to break all OEM's
 > property_contexts files (and AOSP itself).

I get that argument, but it seems like an ugly change to avoid work.

We could version the file in a comment, if no comment assume old and fix it up during the build
with a make warning. They could have a 1to2 (ala python 2 to 3) script that updates it. Warn if they
are using the old syntax and eventually migrate away.

Patch-set: 1
Reviewer: Gerrit User 1013433 <1013433@85c56323-6fa9-3386-8a01-6480fb634889>