Update patch set 1
Patch Set 1:

> I don't think versioning the file is a good idea.  I do like the
 > idea of treating only entries that end with '.' as prefix and the
 > rest as exact match, pending what Jaekyun finds if we make a change
 > like this, but...
 > 
 > A bit most bigger picture, the work that Jaekyun is doing is going
 > to create lots of these exact match properties (~100 from my last
 > check).  This is why I made the previous comment about potential
 > performance issues, as iterating through a list of 100 exact match
 > properties is not a good idea.  It'd be much better to look these
 > up in a hash table or a similar structure that is easily
 > serializable.  We still would need prefix look ups for init.svc.,
 > wrap., ro.boottime., among other prefix properties.
 > 
 > This lead me to the conclusion on an internal bug that we really
 > should have two parallel data structure, one that is the hash table
 > look up for the exact match properties and one that is a trie for
 > the prefix properties (which is separately an improvement over the
 > current look up mechanism).  Ideally these data structures would be
 > created and serialized on the host side for added performance.  In
 > this case, we would have two separate files on the target; one for
 > exact matches and one for prefix matches.
 > 
 > In this case, we can abandon this change, and instead create a
 > separate property_contexts_exactmatch or similarly named file. 
 > Having another file would be a slight downside, but it'd be made up
 > for by having cleaner syntax and the added performance of the above
 > look ups.

Yes there are definitely better data structures that lend itself to
quicker look ups. But if you build these on the target, that would
likely be slower than looking them up in the list. I'm not exactly
sure what your performance requirements are here. Maybe even doing
something like storing len when generating the list could be used
to speed things up. Perhaps it's time to build a binary property
contexts file. If that's the case, that's a perfect time to version
the property specification language and have it passed via the
build system.

With that said, this is a good time to create the syntax what you want it,
even if it does break oem policy, it's not that hard to change it. IMO We
shouldn't kludge the design to avoid doing work.


 > 
 > Lastly, another long term pending item is that the property context
 > reading code really needs to be separated out into a library as
 > both the system properties code in libc and libselinux use it (and
 > unfortunately, it's a copy in libc)

Yes, 100% agreed here.

 > 
 > Thoughts?

Patch-set: 1
Reviewer: Gerrit User 1013433 <1013433@85c56323-6fa9-3386-8a01-6480fb634889>