Update patch set 1
Patch Set 1:

> But if you build these on the target, that would
likely be slower than looking them up in the list.

My ideal goal would be to build them on host, as everything is known at build time.  Treble is the only compounding factor as there are both property contexts as part of /system and /vendor, both of which would need to be searched.  An alternative is to have property service build them in /dev/__properties__/...  In any case, the goal would be to have users mmap() the data structures and use them directly, so no start up costs to any app other than property service.

> At that point, I'd guess that it would cease to be behind the selabel APIs at all, in which case you might as well just take the entire interface and implementation native/specialized to Android, put it in bionic, and drop it from libselinux (both Android and upstream) altogether.  Unless I'm missing something.

Are you saying that this would be a positive thing to do?  I'm all for it if so.  Bionic only uses a subset of the code from libselinux for properties (hence why it's copied; it only needs two static functions for parsing, but uses the data way differently).  Beyond that, the only two clients who ever call selinux_android_prop_context_handle() are property_service and toybox (for `getprop -Z`), so it would be easy to move from libselinux to a new library.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>