secilc: resolve conflicts in expandattribute.
When Android combines multiple cil files from system.img and vendor.img
it's possible to have conflicting expandattribute statements, e.g.
 expandattribute hal_audio true;
 expandattribute hal_audio false;

This change deals with scenario be resolving the value of the
corresponding expandattribute to false. The rationale behind this
override is that "true" is used for reduce run-time lookups, while
"false" is used for tests which must pass.

Bug: 72757373
Test: policy builds

Change-Id: Ia4251ab0f1f4d1b47ef9550f27c73c52b6e88839