Update patch set 2
Patch Set 2:

If we're not careful this could result in gradual performance degradation on devices with diff'd system and non-system partitions.  I'd think the real fix would be to make sure the expand statements are only on the system image.  Otherwise this change looks ok to me.

Patch-set: 2
Reviewer: Gerrit User 1026058 <1026058@85c56323-6fa9-3386-8a01-6480fb634889>