secilc is using bootstrap linker
We are moving the bootstrap bionic libraries and linkers from
/system/lib (or /system/bin) to /system/lib/bootstrap
(or /system/bin/bootstrap). And init bind mounts them to the ordinary
paths. This is causing problem to a few number of processes that are
executed very early - even before the init creates the bind mounts.
Since they need to know the raw paths, they are marked as bootstrap:
true so that their DT_INTERP will point to the raw path under the
bootstrap subdir.

Bug: 120266448
Test: m secilc, readelf -a <path_to_the_built_file> | grep interp shows
[Requesting program interpreter: /system/bin/bootstrap/linker64]

Change-Id: I0a631c78ca00079d5995e3c2addb7ec11bd6d3e3