secilc is using bootstrap Bionic
secilc is a program that is executed very early by init. Mark it as
'bootstrap: true' so that DT_INTERP is set to point to the dynamic
linker for the bootstrap processes. Also RPATH is set to the bootstrap
Bionic path so that this executable can search Bionic libs even before
init makes them visible to the ordinary search path of /system/lib.

Bug: 120266448
Test: m secilc
Inspect DT_INTERP and RPATH with readelf
DT_INTERP: /system/bin/bootstrap/linker64
RPATH: /system/${LIB}/bootstrap

Change-Id: Icb6e2fd37dced18c4cac83fb38e8a28739b34cab