Skip /data/media/* when restorecon recursively /data
The number of user files stored in the /data/media directory
can't be determined. If the number is large and file_contexts
changed after OTA, restorecon /data would need a long time.

The security contexts of files in the /data/media/ directory
are all same and not be changed. So we should skip this directory
to speed up the boot.

Bug: https://issuetracker.google.com/issues/124858106

Test: 1. create a large number files in /sdcard
      2. changed file_contexts and reboot
      3. observer the boot time

Change-Id: I0d8516cb529a143b608487b9f5099652616e28f6
Signed-off-by: liwugang <liwugang@xiaomi.com>