Update patch set 11
Patch Set 11:

(6 comments)

With respect to upstreaming, note that even though libselinux/src/android/* does not exist in upstream libselinux, the setrestoreconlast logic was taken into the upstream selinux_restorecon() function (libselinux/src/selinux_restorecon.c), so potentially the same changes could be made there. I also made a few comments on the code. I'm not sure if Richard is still active. Posting a link to the change on selinux list might be a good idea for wider review, although ultimately submitting an actual patch there would be best (but that won't include the Android-specific bits, so you won't get eyes on those unless you send a link to the change itself or try to upstream the Android-specific bits).

Patch-set: 11