Update patch set 2
Patch Set 2:

Thanks Stephen for sending out the URL where the stem feature was implemented.

I wonder if the stem logic is really needed anymore? The CPU concerns would be relevant when you were trying to relabel on every boot. With the relabeling logic introduced in this change, the amount of time you'd spend with the additional CPU work would be drastically decreased because you're no longer walking every file.

IIUC, the stem logic was introduced to make relabeling faster, but this change does it in a better way. Maybe we should get rid of the stem logic entirely?

Patch-set: 2