Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Wait, does this permission apply to restorecon as well? If an under privileged caller supply this flag, could a given directory get the new context but still have an old hash value?

Yes, this applies to restorecon as well, if applied to a directory. (I believe the code also sets the security.sehash attribute then).

Regardless of whether the underprivileged caller specifies the flag or not, the directory will get the updated context but will have the old hash value. The setxattr call is unsuccessful if the caller doesn't specify the flag, and will be skipped if the caller does specify the flag.

In the case that this change is intended to address, it covers relabeling of subdirectories of /data/app. When init hits this directory, if the /data/app sub-regexes haven't changed, the recursion will stop here, without ever being able to see the hashes of the subdirectories.

It seems like there should be a mechanism to set the hash that doesn't depend on the caller having CAP_SYS_ADMIN, but the kernel doesn't support that right now. Instead of using the "security.*" namespace, one could imagine using the "user.*" namespace which doesn't require any special permissions. But then anyone could adjust the hash of their directories, which I'm not sure is an improvement.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0