Set apex_available property
libselinux has been  available to the platform and every APEXes via the
hand-written whitelist in build/soong/apex/apex.go. Trying to remove the
whitelist by adding apex_available property to the Android.bp of the
library.

This is safe because the interface between libselinux and the kernel is
guaranteed to be stable (i.e. libselinux is considered as portable)

Bug: 151053366
Bug: 150999716
Test: m
Change-Id: Iaad62a51683d7cb1b3e0b1f778f84b4dd843cd80