Optimize the calculation of security.sehash
The hash of each directory should be determined by the file contexts of
the current directory and subdirectories, but the existing logic also
includes the ancestor directories. The first optimization is to exclude
them.

If the current directory has corresponding file contexts and subdirectories
have not, subdirectories will use the current direcorty's. There is no need
to calculate the hash for the subdirectories. It will save time, espacially
for user data(/data/media/0/). The second optimization is to ignore the
subdirectories.

Example:
/data/(.*)?           u:object_r:system_data_file:s0
/data/backup(/.*)?    u:object_r:backup_data_file:s0

If the file context of "/data/(.*)?" changes and "/data/backup(/.*)?" does not
change, directory(/data/backup) and the subdirectories will restorecon because of
hash changed. But actually there is no need the restorecon.

Test: compile
Signed-off-by: liwugang <liwugang@xiaomi.com>
Change-Id: I95de63235ab6f8f92a73d360bf1ec04b3a50d619