Revert "Fix build after upstream merge"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84b3e97a-686c-4f32-9bd6-b198d82ba447). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib3b48e89be5e996ebeb5daff01be9ffa79b89c06