Revert "Integrate isIsolatedComputeApp selection logics for seco..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48185cee-b93d-48af-b2d0-f8ce78e086c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17f94f913dd8205ba23d0b292c5567e9e341f61f