Revert "Integrate isIsolatedComputeApp selection logics for seco..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9153ddaf-5860-41f0-98b4-8d9566dd207f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If2e7b36434567d4b9cdf6295ad178cd814425807