Revert "Integrate isIsolatedComputeApp selection logics for seco..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc95d6b4-1f07-49c5-a3c9-f6dd0b41d84b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72b7ace272b19621437c74aec878ad9ef08fd818