Revert "Integrate isIsolatedComputeApp selection logics for seco..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71a87e25-8a44-445e-a0af-79f7559c2a5f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If39f5641dc9bf11cdfaf40781c2c498a891a4b11