Revert "Integrate isIsolatedComputeApp selection logics for seco..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ba4a4c6-de7e-4f77-8643-d8b602018b5f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I899ccadc14258e768a9e887c209773988978fdde