Add callback to report user avc to logd.auditd
When a denial occurs in userspace (e.g., in servicemanager), it is
logged directly. This bypasses the processing in auditd (e.g., to attach a
bug based on bug_map). This change introduces a new callback
selinux_log_netlink_callback which will report any avc message to audit
via netlink.

Two functions are included to open and close the netlink socket.

All the changes are only exported to the platform (see
LIBSELINUX_PLATFORM).

Bug: 235163405
Test: Trigger denial within servicemanager, using this new callback.
Change-Id: I7e90460f3982bdd66f4e1bf70eb836739e1dc811