Consider /data/data as an app_data directory
When an OTA is performed, if an entry for /data/data/* was added or
removed from file_contexts (which should not have existed in the first
place), restorecon will try to process /data/data subdirectories. It
should have been filtered out by is_app_data_path, but the base
directory was not matched (/data/data vs /data/data/). Explicitly
match on the base directory.

Bug: 317296680
Test: build, flash, boot, add an entry to file_contexts for /data/data,
    rebuild, and flash /system. At the first boot, restorecon is
    correctly executed. Subsequent boots skip restorecon.
Change-Id: I992bfdad7429208b221e9ae3718c137e744dee9e