Revert "Add selinux_android_tee_service_context_handle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/23442cea-6e1c-43cf-879f-44fd9af2e737). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379196597
Change-Id: I63f07339d8660d532fcf649e7dd9380769669c40
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com