Revert "Export getprevcon"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d7deaf5-9333-485e-a864-7e07a2eb8eea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379159866
Change-Id: Idfc0a0a8314f618e2bac24534f9572547b55ce82
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com