Revert "Add selinux_android_tee_service_context_handle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d7deaf5-9333-485e-a864-7e07a2eb8eea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379159866
Change-Id: I01c9e8e07c7dac0bf2742fd88b3cfae1290f58dc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com