Revert "Export getprevcon"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d70ea2a3-d1ba-454a-981f-e47254a4ffb5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b0a0f8a906e8437fcbb15f7bb2983b10e67a000
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com